pyproject.toml
[project]
name = "llm-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "openai>=2.13.0",
    "huggingface-hub>=0.23.0",
    "transformers>=4.40.0",
    "torch>=2.1.0",
    "accelerate>=0.30.0",
    "sentencepiece>=0.1.99",
    "pandas>=2.3.3",
    "bitsandbytes>=0.49.0",
    "tiktoken>=0.12.0",
    "einops>=0.8.1",
    "pytest>=9.0.2",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
    "google-genai>=1.57.0",
    "matplotlib>=3.10.8",
    "streamlit>=1.53.0",
    "scipy>=1.17.0",
    "ipywidgets>=8.1.8",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "tdqm>=0.0.1",
    "transformers>=4.57.3",
]
